pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "strava-agent"
version = "0.1.0"
description = "Multi-Agent Solution for Analysing Strava Data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-telegram-bot>=20.0.0",
    "langchain>=0.2.0",
    "langchain-google-genai>=0.1.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0", 
    "ruff>=0.1.0"
]
